+++ internal/cmd/plugin/pkg/logs/cluster_logs.go
