+++ internal/cmd/plugin/pkg/logs/cluster_logs_test.go
