+++ internal/cmd/plugin/pkg/logs/logs_test.go
